refactor(inventory_service): Reorganize folder structure
