+++ services/inventory-service/requests.http
@@ -13,18 +13,18 @@ content-type: application/json
     "description": "Test ürünü",
     "category": "Meyve",
     "unitOfMeasure": "adet",
-    "safetyStock": 10.0,
-    "reorderPoint": 5.0,
+    "safetyStock": 5.0,
+    "reorderPoint": 10.0,
     "currentPrice": 99.99
 }
 
 ### Get Product By ID
 
-GET http://localhost:8000/api/v1/products/2
+GET http://localhost:8000/api/v1/products/1
 
 ### Update Product
 
-PUT http://localhost:8000/api/v1/products/5
+PUT http://localhost:8000/api/v1/products/1
 content-type: application/json
 
 {
@@ -35,4 +35,4 @@ content-type: application/json
 
 ### Delete Product
 
-DELETE http://localhost:8000/api/v1/products/4
+DELETE http://localhost:8000/api/v1/products/1
