+++ services/inventory-service/src/main/java/com/petek/inventory_service/dto/ProductRequest.java
@@ -1,4 +1,4 @@
-package com.petek.inventory_service.product;
+package com.petek.inventory_service.dto;
 
 import java.math.BigDecimal;
