+++ services/inventory-service/src/main/java/com/petek/inventory_service/entity/Customer.java
@@ -1,7 +1,9 @@
-package com.petek.inventory_service.supplier;
+package com.petek.inventory_service.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -19,15 +21,24 @@ import lombok.Setter;
 @Getter
 @Setter
 @Entity
-@Table(name = "suppliers")
-public class Supplier {
+@Table(name = "customers")
+public class Customer {
+    
+    public enum CustomerSegment {
+        INDIVIDUAL, SME, CORPORATE, ENTERPRISE, OTHER
+    }
+    
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "supplier_id")
-    private Long supplierId;
+    @Column(name = "customer_id")
+    private Long customerId;
     
-    @Column(name = "supplier_name", nullable = false, length = 200)
-    private String supplierName;
+    @Column(name = "customer_name", nullable = false, length = 200)
+    private String customerName;
+    
+    @Enumerated(EnumType.STRING)
+    @Column(name = "customer_segment")
+    private CustomerSegment customerSegment;
     
     @Column(length = 100)
     private String email;
