+++ services/inventory-service/src/main/java/com/petek/inventory_service/entity/Product.java
@@ -1,4 +1,4 @@
-package com.petek.inventory_service.product;
+package com.petek.inventory_service.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
