+++ services/inventory-service/src/main/java/com/petek/inventory_service/repository/ProductRepository.java
@@ -1,6 +1,8 @@
-package com.petek.inventory_service.product;
+package com.petek.inventory_service.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.petek.inventory_service.entity.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Long>{
